trek-api/requirements.txt
fastapi
uvicorn
langgraph==0.3.5
langgraph-checkpoint-postgres
langchain-groq
langchain-core
psycopg-pool
psycopg[binary]
supabase
groq
tavily-python
httpx
python-dotenv
firecrawl-py
